requirements.txt
aiohttp=3.7.4
requests=2.28.1
loguru=0.5.3
openai-whisper
pytorch
torchvision
torchaudio
pytorch-cuda=11.7
chardet
ffmpeg-python
streamlink
aiohttp
cudatoolkit